Restoring the hours display on Document Services
This restores the call from postHead to locationHours, and instead
removes (comments out) all the calculation of the old hours
infrastructure. Instead, locationHours is simply static markup that then
gets filled by the javascript call.

As a to-do, I'm not certain that locationHours is still needed - or at
least can be much shorter, with all PHP code now commented out.